requirements.txt
aiosql==9.3
psycopg==3.1.17
psycopg-binary==3.1.17
PyYAML==6.0.1
typing_extensions==4.9.0
